requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
python-dotenv
google-genai
pypdf
python-docx
jsonschema
sqlalchemy
psycopg2-binary
celery
redis
boto3
passlib
python-jose
email-validator
requests
beautifulsoup4
httpx